Requests for moderating and user view are in separated apps. Access restrictions added

+++ DjangoRestApi/moderate/views.py
@@ -7,6 +7,7 @@ from rest_framework import status
 from storage.models import Upload
 from thematic_pages.models import ThematicPage
 from storage.serializers import UploadSerializer
+from thematic_pages.serializers import ThematicPageSerializer
 from rest_framework.decorators import api_view
 
 """
@@ -15,7 +16,22 @@ Methods	    Urls                                Actions
 GET         /api/moderate/thematic_page/:id     get Upload items to moderate for Thematic Page with defined id     
 PUT         /api/moderate/upload/:id/approve    approve upload with given id
 PUT         /api/moderate/upload/:id/reject     reject upload with given id
+GET         /api/moderate/thematic_pages_list   get ThematicPages to moderate
 """
+
+
+@api_view(['GET'])
+def moderate_pages_list(request):
+    # get pages available for moderating to user
+    thematic_pages = request.user.can_moderate
+
+    if request.method == 'GET':
+        # serialize pages
+        thematic_page_serializer = ThematicPageSerializer(thematic_pages, many=True)
+        # return serialized page
+        return JsonResponse(thematic_page_serializer.data, safe=False)
+
+
 @api_view(['GET'])
 def moderate_page(request, pk):
     # Get page for moderating
@@ -28,10 +44,13 @@ def moderate_page(request, pk):
 
     # GET request
     if request.method == 'GET':
-        # Serialize specified upload
-        upload_serializer = UploadSerializer(upload, many=True)
-        # Return serialized upload
-        return JsonResponse(upload_serializer.data, safe=False)
+        if page in request.user.can_moderate:
+            # Serialize specified upload
+            upload_serializer = UploadSerializer(upload, many=True)
+            # Return serialized upload
+            return JsonResponse(upload_serializer.data, safe=False)
+        else:
+            return JsonResponse({'message': "User can't moderate this page"}, status=status.HTTP_403_FORBIDDEN)
 
 
 @api_view(['PUT'])
@@ -43,9 +62,12 @@ def approve_upload(request, pk):
         return JsonResponse({'message': 'The upload does not exist'}, status=status.HTTP_404_NOT_FOUND)
     # PUT request
     if request.method == 'PUT':
-        upload.status = 1
-        # Return data with status = 1
-        return JsonResponse({'message': 'The upload is now approved!'})
+        if upload.thematic_page in request.user.can_moderate:
+            upload.status = 1
+            # Return data with status = 1
+            return JsonResponse({'message': 'The upload is now approved!'})
+        else:
+            return JsonResponse({'message': "User can't moderate this upload"}, status=status.HTTP_403_FORBIDDEN)
 
 
 @api_view(['PUT'])
@@ -57,7 +79,10 @@ def reject_upload(request, pk):
         return JsonResponse({'message': 'The upload does not exist'}, status=status.HTTP_404_NOT_FOUND)
     # PUT request
     if request.method == 'PUT':
-        # delete upload
-        upload.delete()
-        # Notify that item is now rejected and deleted
-        return JsonResponse({'message': 'Upload rejected and deleted!'}, status=status.HTTP_204_NO_CONTENT)
+        if upload.thematic_page in request.user.can_moderate:
+            # delete upload
+            upload.delete()
+            # Notify that item is now rejected and deleted
+            return JsonResponse({'message': 'Upload rejected and deleted!'}, status=status.HTTP_204_NO_CONTENT)
+        else:
+            return JsonResponse({'message': "User can't moderate this upload"}, status=status.HTTP_403_FORBIDDEN)
